+++ react/Payment.tsx
@@ -4,9 +4,10 @@ import { Router, routes } from 'vtex.checkout-container'
 import { AvailableAccount, PaymentSystem } from 'vtex.checkout-graphql'
 
 import CreditCard from './CreditCard'
-import Installments from './Installments'
 import PaymentList from './PaymentList'
 import { PaymentStage } from './enums/PaymentEnums'
+import InstallmentsModal from './components/InstallmentsModal'
+import ExtraData from './components/ExtraData'
 
 const { useHistory } = Router
 
@@ -15,7 +16,6 @@ const Payment: React.FC = () => {
   const [cardType, setCardType] = useState<CardType>('new')
   const {
     setPaymentField,
-    cardLastDigits,
     setCardLastDigits,
     value,
     referenceValue,
@@ -23,23 +23,25 @@ const Payment: React.FC = () => {
   } = useOrderPayment()
   const history = useHistory()
 
-  const goToCardForm = () => {
+  const [installmentsModalOpen, setInstallmentsModalOpen] = useState(false)
+
+  const handleCardFormCompleted = () => {
+    setCardFormFilled(true)
+    setInstallmentsModalOpen(true)
+  }
+
+  const handleDeselectPayment = () => {
     setCardFormFilled(false)
     setStage(PaymentStage.CARD_FORM)
   }
 
-  const goToInstallments = () => {
-    setCardFormFilled(true)
-    setStage(PaymentStage.INSTALLMENTS)
-  }
-
-  const goToPaymentList = () => {
+  const handleChangePaymentMethod = () => {
     setStage(PaymentStage.PAYMENT_LIST)
   }
 
   const handleNewCreditCard = () => {
     setCardType('new')
-    goToCardForm()
+    handleDeselectPayment()
   }
 
   const handleSavedCreditCard = async (payment: AvailableAccount) => {
@@ -50,14 +52,15 @@ const Payment: React.FC = () => {
       accountId: payment.accountId,
       bin: payment.bin,
     })
-    goToCardForm()
+    handleDeselectPayment()
   }
 
   const handleInstallmentSelected = async (installment: number) => {
+    setInstallmentsModalOpen(false)
     await setPaymentField({
       installments: installment,
     })
-    history.push(routes.REVIEW)
+    setStage(PaymentStage.EXTRA_DATA)
   }
 
   const handleBankInvoiceSelect = async (payment: PaymentSystem) => {
@@ -71,12 +74,16 @@ const Payment: React.FC = () => {
     history.push(routes.REVIEW)
   }
 
+  const handleExtraDataSubmit = () => {
+    history.push(routes.REVIEW)
+  }
+
   return (
     <>
       <div className={stage === PaymentStage.CARD_FORM ? '' : 'dn'}>
         <CreditCard
-          onCardFormCompleted={goToInstallments}
-          onChangePaymentMethod={goToPaymentList}
+          onCardFormCompleted={handleCardFormCompleted}
+          onChangePaymentMethod={handleChangePaymentMethod}
           cardType={cardType}
           key={cardType}
         />
@@ -87,13 +94,19 @@ const Payment: React.FC = () => {
           onSavedCreditCard={handleSavedCreditCard}
           onBankInvoiceSelect={handleBankInvoiceSelect}
         />
-      ) : stage === PaymentStage.INSTALLMENTS ? (
-        <Installments
-          onInstallmentSelected={handleInstallmentSelected}
-          onBackToCardForm={goToCardForm}
-          cardLastDigits={cardLastDigits}
+      ) : stage === PaymentStage.EXTRA_DATA ? (
+        <ExtraData
+          onDeselectPayment={handleChangePaymentMethod}
+          cardType={cardType}
+          onSubmit={handleExtraDataSubmit}
+          onChangeInstallments={() => setInstallmentsModalOpen(true)}
         />
       ) : null}
+      <InstallmentsModal
+        isOpen={installmentsModalOpen}
+        onInstallmentSelected={handleInstallmentSelected}
+        onClose={() => setInstallmentsModalOpen(false)}
+      />
     </>
   )
 }